#8 [DE] Added `setMap` into `Mode` component to update map as well
